dissolve state machine
