--- mcr-core/mcr_meeting/app/state_machine/visio_meeting_state_machine.py
@@ -1,97 +1,0 @@
-from statemachine import State, StateMachine
-
-from mcr_meeting.app.models import Meeting, MeetingStatus
-from mcr_meeting.app.statemachine_actions.meeting_actions import (
-    after_transition_handler,
-    update_status_handler,
-)
-
-
-# Tip: Run `make generate-graphs` to automatically generate all state machine diagrams
-# and save them in the `mcr_meeting/app/state_machine/graph/` directory.
-class VisioMeetingStateMachine(StateMachine):
-    def __init__(self, meeting: Meeting | None = None):
-        self.meeting: Meeting | None = meeting
-        if meeting is None:
-            super().__init__()
-        else:
-            super().__init__(model=meeting, start_value=meeting.status)
-
-    # -------------------------------------------------------------------------
-    # STATES
-    # -------------------------------------------------------------------------
-    NONE = State(MeetingStatus.NONE, initial=True)
-    CAPTURE_PENDING = State(MeetingStatus.CAPTURE_PENDING)
-    CAPTURE_BOT_IS_CONNECTING = State(MeetingStatus.CAPTURE_BOT_IS_CONNECTING)
-    CAPTURE_BOT_CONNECTION_FAILED = State(MeetingStatus.CAPTURE_BOT_CONNECTION_FAILED)
-    CAPTURE_IN_PROGRESS = State(MeetingStatus.CAPTURE_IN_PROGRESS)
-    CAPTURE_FAILED = State(MeetingStatus.CAPTURE_FAILED)
-    CAPTURE_DONE = State(MeetingStatus.CAPTURE_DONE)
-    TRANSCRIPTION_PENDING = State(MeetingStatus.TRANSCRIPTION_PENDING)
-    TRANSCRIPTION_IN_PROGRESS = State(MeetingStatus.TRANSCRIPTION_IN_PROGRESS)
-    TRANSCRIPTION_DONE = State(MeetingStatus.TRANSCRIPTION_DONE)
-    TRANSCRIPTION_FAILED = State(MeetingStatus.TRANSCRIPTION_FAILED)
-    REPORT_PENDING = State(MeetingStatus.REPORT_PENDING)
-    REPORT_DONE = State(MeetingStatus.REPORT_DONE)
-    REPORT_FAILED = State(MeetingStatus.REPORT_FAILED)
-    DELETED = State(MeetingStatus.DELETED, final=True)
-
-    # -------------------------------------------------------------------------
-    # TRANSITIONS / EVENTS
-    # -------------------------------------------------------------------------
-    INIT_CAPTURE = NONE.to(CAPTURE_PENDING) | CAPTURE_BOT_CONNECTION_FAILED.to(
-        CAPTURE_PENDING
-    )
-    START_CAPTURE_BOT = CAPTURE_BOT_IS_CONNECTING.to(CAPTURE_IN_PROGRESS)
-    FAIL_CAPTURE_BOT = CAPTURE_BOT_IS_CONNECTING.to(CAPTURE_BOT_CONNECTION_FAILED)
-    START_CAPTURE = CAPTURE_PENDING.to(CAPTURE_BOT_IS_CONNECTING)
-    COMPLETE_CAPTURE = CAPTURE_IN_PROGRESS.to(CAPTURE_DONE)
-    FAIL_CAPTURE = CAPTURE_IN_PROGRESS.to(CAPTURE_FAILED)
-    INIT_TRANSCRIPTION = (
-        CAPTURE_DONE.to(TRANSCRIPTION_PENDING)
-        | CAPTURE_FAILED.to(TRANSCRIPTION_PENDING)
-        | TRANSCRIPTION_FAILED.to(TRANSCRIPTION_PENDING)
-    )
-    START_TRANSCRIPTION = TRANSCRIPTION_PENDING.to(TRANSCRIPTION_IN_PROGRESS)
-    COMPLETE_TRANSCRIPTION = TRANSCRIPTION_IN_PROGRESS.to(TRANSCRIPTION_DONE)
-    # We changed from an internal transition to a external transition because
-    # it is more visual on the generated graphs. At the time of change, the functionalities
-    # are the same because internal transition only skip on_enter_* and on_exit_*
-    # But it doesn't after after_* and before_* transitions actions
-    UPDATE_TRANSCRIPTION = TRANSCRIPTION_DONE.to.itself() | REPORT_DONE.to(  # type: ignore[no-untyped-call]
-        TRANSCRIPTION_DONE
-    )
-    FAIL_TRANSCRIPTION = TRANSCRIPTION_PENDING.to(
-        TRANSCRIPTION_FAILED
-    ) | TRANSCRIPTION_IN_PROGRESS.to(TRANSCRIPTION_FAILED)
-    START_REPORT = TRANSCRIPTION_DONE.to(REPORT_PENDING)
-    COMPLETE_REPORT = REPORT_PENDING.to(REPORT_DONE)
-    FAIL_REPORT = REPORT_PENDING.to(REPORT_FAILED)
-    RESET_REPORT = REPORT_DONE.to(TRANSCRIPTION_DONE) | REPORT_FAILED.to(
-        TRANSCRIPTION_DONE
-    )
-    # Ignore mypy warning: from_.any() is dynamic DSL, not typed
-    DELETE = DELETED.from_.any()  # type: ignore
-
-    # -------------------------------------------------------------------------
-    # AFTER HOOKS (SIDE EFFECTS)
-    # -------------------------------------------------------------------------
-    def after_START_CAPTURE(self) -> None:
-        if self.meeting is None:
-            return
-        update_status_handler(self.meeting, self.current_state_value)
-
-    def after_FAIL_CAPTURE(self) -> None:
-        if self.meeting is None:
-            return
-        update_status_handler(self.meeting, self.current_state_value)
-
-    def after_DELETE(self) -> None:
-        if self.meeting is None:
-            return
-        update_status_handler(self.meeting, self.current_state_value)
-
-    def after_transition(self) -> None:
-        if self.meeting is None:
-            return
-        after_transition_handler(self.meeting.id, self.current_state_value)
