+++ mcr-core/scripts/generate_state_machine_graph.py
@@ -2,43 +2,42 @@ import importlib
 import inspect
 import os
 import sys
-from typing import Union, List, Set
 
+import pydot
 from loguru import logger
 from statemachine import StateMachine
 from statemachine.contrib.diagram import DotGraphMachine
-import pydot
 
 PROJECT_ROOT = os.path.abspath(os.path.join(os.path.dirname(__file__), ".."))
 if PROJECT_ROOT not in sys.path:
     sys.path.insert(0, PROJECT_ROOT)
 
+# Module holding the meeting state machines (validators, I/O-less).
+STATE_MACHINE_MODULE = "mcr_meeting.app.domain.meeting_state_machine"
+
+# Terminal states drawn muted and pushed to the bottom of the graph.
 GLOBAL_STATES = ["DELETED"]
+
+OUTPUT_DIR = os.path.join(PROJECT_ROOT, "docs", "state_machine")
 
 
 def create_graph_with_hidden_states(
-    machine: Union[type[StateMachine], StateMachine],
-    hidden_states: Union[List[str], Set[str]] = None,
+    machine: type[StateMachine] | StateMachine,
+    hidden_states: list[str] | set[str] | None = None,
 ) -> pydot.Dot:
-    """
-    Generate a state machine diagram with specific states visually minimized
-    and positioned as sink nodes.
-    """
+    """Generate a diagram with the given states visually minimized and positioned
+    as sink nodes."""
     hidden_states = set(hidden_states or [])
-    graph_generator = DotGraphMachine(machine)
-    dot_graph = graph_generator()
+    dot_graph = DotGraphMachine(machine)()
 
     if not hidden_states:
         return dot_graph
 
-    # hidden states subgraph
     sink_subgraph = pydot.Subgraph(rank="sink")
 
     for node in dot_graph.get_nodes():
         node_name = node.get_name().strip('"')
-
         if node_name in hidden_states:
-            # muted styles
             node.set_style("filled")
             node.set_fillcolor("#eeeeee")
             node.set_color("gray")
@@ -50,7 +49,6 @@ def create_graph_with_hidden_states(
     for edge in dot_graph.get_edges():
         source = edge.get_source().strip('"')
         dest = edge.get_destination().strip('"')
-
         if source in hidden_states or dest in hidden_states:
             edge.set_style("dotted")
             edge.set_color("gray")
@@ -59,58 +57,36 @@ def create_graph_with_hidden_states(
     return dot_graph
 
 
-def main():
+def main() -> None:
     _import_env()
 
-    base_dir = os.path.join("mcr_meeting", "app", "state_machine")
-    graph_dir = os.path.join(base_dir, "graph")
+    os.makedirs(OUTPUT_DIR, exist_ok=True)
 
-    if not os.path.exists(graph_dir):
-        os.makedirs(graph_dir)
+    module = importlib.import_module(STATE_MACHINE_MODULE)
 
     total_graphs = 0
+    for name, obj in inspect.getmembers(module, inspect.isclass):
+        is_local_state_machine = (
+            issubclass(obj, StateMachine)
+            and obj is not StateMachine
+            and obj.__module__ == module.__name__
+        )
+        if not is_local_state_machine:
+            continue
 
-    for root, _, files in os.walk(base_dir):
-        for file in files:
-            if file.endswith(".py") and file != "__init__.py":
-                file_path = os.path.join(root, file)
-                rel_path = os.path.relpath(file_path, start=os.getcwd())
-                module_path = rel_path[:-3].replace(os.sep, ".")
+        try:
+            graph = create_graph_with_hidden_states(obj(), hidden_states=GLOBAL_STATES)
+            output_file = os.path.join(OUTPUT_DIR, f"{name}.png")
+            graph.write_png(path=output_file)
+            logger.info("Graph generated successfully for {}: {}", name, output_file)
+            total_graphs += 1
+        except Exception as e:
+            logger.error("Error generating graph for {}: {}", name, e)
 
-                try:
-                    module = importlib.import_module(module_path)
-                except Exception as e:
-                    logger.error("Error importing module {}: {}", module_path, e)
-                    continue
-
-                for name, obj in inspect.getmembers(module, inspect.isclass):
-                    if issubclass(obj, StateMachine) and obj is not StateMachine:
-                        try:
-                            sm = obj()
-                            graph = DotGraphMachine(sm)
-                            graph = create_graph_with_hidden_states(
-                                sm, hidden_states=GLOBAL_STATES
-                            )
-                            output_file = os.path.join(graph_dir, f"{name}.png")
-                            graph.write_png(path=output_file)
-                            logger.info(
-                                "Graph generated successfully for {}: {}",
-                                name,
-                                output_file,
-                            )
-                            total_graphs += 1
-                        except Exception as e:
-                            logger.error(
-                                "Error generating graph for {} in {}: {}",
-                                name,
-                                module_path,
-                                e,
-                            )
-
-    logger.info("Summary: {} graph(s) generated.", total_graphs)
+    logger.info("Summary: {} graph(s) generated in {}.", total_graphs, OUTPUT_DIR)
 
 
-def _import_env():
+def _import_env() -> None:
     os.environ.setdefault("DISABLE_DB", "1")
     os.environ.setdefault("POSTGRES_USER", "dummy")
     os.environ.setdefault("POSTGRES_PASSWORD", "dummy")
